+++ smithy-swift-codegen/src/main/kotlin/software/amazon/smithy/swift/codegen/integration/HttpProtocolCustomizable.kt
@@ -13,8 +13,6 @@ import software.amazon.smithy.protocoltests.traits.HttpRequestTestCase
 import software.amazon.smithy.swift.codegen.SwiftWriter
 
 interface HttpProtocolCustomizable {
-    //fun renderMiddlewares(ctx: ProtocolGenerator.GenerationContext, writer: SwiftWriter, op: OperationShape, operationStackName: String)
-
     fun renderInternals(ctx: ProtocolGenerator.GenerationContext) {
         // Default implementation is no-op
     }
@@ -57,14 +55,4 @@ interface HttpProtocolCustomizable {
     fun alwaysHasHttpBody(): Boolean {
         return false
     }
-
-    /**
-     * Get all of the middleware that should be installed into the operation's middleware stack (`SdkOperationExecution`)
-     * This is the function that protocol client generators should invoke to get the fully resolved set of middleware
-     * to be rendered (i.e. after integrations have had a chance to intercept). The default set of middleware for
-     * a protocol can be overridden by [baseMiddlewares].
-     */
-    //fun operationMiddlewares(ctx: ProtocolGenerator.GenerationContext, op: OperationShape): List<OperationMiddlewareRenderable>
-
-    //fun baseMiddlewares(ctx: ProtocolGenerator.GenerationContext, op: OperationShape): List<OperationMiddlewareRenderable>
 }
