+++ smithy-swift-codegen/src/main/kotlin/software/amazon/smithy/swift/codegen/integration/OperationMiddlewareRenderable.kt
@@ -8,6 +8,8 @@ package software.amazon.smithy.swift.codegen.integration
 import software.amazon.smithy.model.shapes.OperationShape
 import software.amazon.smithy.model.shapes.ServiceShape
 import software.amazon.smithy.swift.codegen.SwiftWriter
+import software.amazon.smithy.swift.codegen.middleware.MiddlewarePosition
+import software.amazon.smithy.swift.codegen.middleware.MiddlewareStep
 
 /**
  * Interface that allows middleware to be registered and configured with the generated protocol client
