+++ smithy-swift-codegen/src/main/kotlin/software/amazon/smithy/swift/codegen/integration/middlewares/ContentMD5Middleware.kt
@@ -3,8 +3,8 @@ package software.amazon.smithy.swift.codegen.integration.middlewares
 import software.amazon.smithy.model.shapes.OperationShape
 import software.amazon.smithy.swift.codegen.ClientRuntimeTypes
 import software.amazon.smithy.swift.codegen.SwiftWriter
-import software.amazon.smithy.swift.codegen.integration.MiddlewarePosition
-import software.amazon.smithy.swift.codegen.integration.MiddlewareStep
+import software.amazon.smithy.swift.codegen.middleware.MiddlewarePosition
+import software.amazon.smithy.swift.codegen.middleware.MiddlewareStep
 import software.amazon.smithy.swift.codegen.model.capitalizedName
 
 class ContentMD5Middleware {
